toolkit/components/pdfjs/test/test_pdf_file_in_iframe.html

<!DOCTYPE html>
<html>

<head>
  <meta charset="utf-8">
  <title>PDFjs: Load a PDF in an iframe.</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<body>
  <pre id="test">
        <script type="text/javascript">
          "use strict";

          async function test() {
            await SpecialPowers.pushPrefEnv({
              "set": [
                ["pdfjs.disabled", false],
                ["pdfjs.enableFloatingToolbar", false],
              ],
            });

            const iframe = document.createElement("iframe");
            document.body.append(iframe);
            iframe.src = "file_pdfjs_test.pdf";
            iframe.onload = async () => {
              const hasViewerContainer = await SpecialPowers.spawn(
                iframe.contentWindow,
                [],
                () => !!this.content.document.getElementById("viewerContainer")
              );
              ok(
                hasViewerContainer,
                "The iframe with a pdf must have a `viewerContainer`"
              );
              SimpleTest.finish();
            };
          }

          window.onload = () => {
            SimpleTest.waitForExplicitFinish();
            test();
          };
        </script>
    </pre>
</body>

</html>
